requirements.txt
# FHIR R4-compatible libraries
fhir.resources==6.1.0

# Required Pydantic version for fhir.resources 6.1.0
pydantic<2.0

# Optional: requests library for interacting with FHIR servers
requests>=2.20.0,<3.0.0

# Add fhirpath-py for FHIRPath expressions
fhirpathpy

# Add Textual Support for a simple front end framework
textual==0.44.1